Comments and small refactoring in Config
Adds missing comments to classes and small refactoring to tidy up the
output dir function in Config
