+++ src/main/java/org/example/Config.java
@@ -24,17 +24,27 @@ public class Config {
     private static final String DEFAULT_PURPOSE = "default";
     
     // ---- Required Parameters ----
+    /** Filepath to read records csv from */
     private String recordsFilepath;
+    /** Type of record, e.g BIRTH, MARRIAGE, DEATH */
     private String recordType;
+    /** Record format, e.g DS or TD */
     private String recordFormat;
 
     // ---- Optional Parameters ----
     // TODO: Make this some form of list to perform multiple operations
+    /** Analysis operation to perform on record-data */
     private String analysisType = DEFAULT_ANALYSIS_TYPE;
+    /** Results directory to store output in */
     private String resultsDir = DEFAULT_RESULTS_DIR;
+    /** Purpose of run (used for categorising results in results dir) */
     private String purpose = DEFAULT_PURPOSE;
 
     // ---- Set automatically at runtime ----
+    /**
+     * Directory to output results to for this run. Takes structure
+     * <resultsDir/<purpose>/<timestamp>
+     */
     private String outputDir;
 
     public Config(String recordsFilepath, String recordType, String recordFormat) {
@@ -54,13 +64,19 @@ public class Config {
         this.outputDir = generateOutputFilepath();
     }
 
+    /**
+     * Generates a Map of key-function pairs to use to process different keys
+     * read-in from the config file.
+     * 
+     * @return Map of <key, processor-function> pairs for the config file
+     */
     private Map<String, BiConsumer<Config, String>> getProcessors() {
         Map<String, BiConsumer<Config, String>> processors = 
             new HashMap<String, BiConsumer<Config, String>>();
 
         processors.put(Constants.RECORDS_FILEPATH_KEY, Config::setRecordsFilepath);
-        processors.put(Constants.RECORDS_FORMAT_KEY, Config::setRecordFormat);
-        processors.put(Constants.RECORDS_TYPE_KEY, Config::setRecordType);
+        processors.put(Constants.RECORD_FORMAT_KEY, Config::setRecordFormat);
+        processors.put(Constants.RECORD_TYPE_KEY, Config::setRecordType);
         processors.put(Constants.ANALYSIS_TYPE_KEY, Config::setAnalysisType);
         processors.put(Constants.RESULTS_FILEPATH_KEY, Config::setResultsDir);
         processors.put(Constants.PURPOSE_KEY, Config::setPurpose);
@@ -69,6 +85,14 @@ public class Config {
         
     }
     
+    /**
+     * Processes lines read-in from a config-file, using processor functions
+     * specified for each key.
+     * 
+     * @param lines
+     * @param processors Map of <key, processor-function> pairs for the config
+     *                   file
+     */
     private void processArgs(List<String> lines, Map<String, BiConsumer<Config, String>> processors) {
         String[] parts;
         String key;
@@ -83,25 +107,31 @@ public class Config {
             }
 
             key = parts[0].trim();
-            value = String.join("=", Arrays.copyOfRange(parts, 1, parts.length)).trim();
-
             if ((processor = processors.get(key)) == null) {
                 throw new IllegalArgumentException("`" + key + "` is not a recognised option");
             }
 
+            // Incase values string contains '='
+            value = String.join("=", Arrays.copyOfRange(parts, 1, parts.length)).trim();
             processor.accept(this, value);
         }
     }
 
-    private void validateArgs() {
+    /**
+     * Validate arguments provided in configuration file, checks validity and
+     * that all required options are set.
+     * 
+     * @throws IllegalArgumentException
+     */
+    private void validateArgs() throws IllegalArgumentException {
         if (recordsFilepath == null) {
             throw new IllegalArgumentException("`" + Constants.RECORDS_FILEPATH_KEY + "` is required");
         }
         if (recordType == null) {
-            throw new IllegalArgumentException("`" + Constants.RECORDS_FORMAT_KEY + "` is required");
+            throw new IllegalArgumentException("`" + Constants.RECORD_FORMAT_KEY + "` is required");
         }
         if (recordFormat == null) {
-            throw new IllegalArgumentException("`" + Constants.RECORDS_TYPE_KEY + "` is required");
+            throw new IllegalArgumentException("`" + Constants.RECORD_TYPE_KEY + "` is required");
         }
         if (!Arrays.asList(Constants.TYPES).contains(recordType)) {
             throw new IllegalArgumentException("Unrecognised record type: `" + recordType + "`");
@@ -114,16 +144,21 @@ public class Config {
         }
     }
 
+    /**
+     * Generates filepath to use for outputs
+     * (<resultsDir/<purpose>/<timestamp>).
+     * 
+     * @return output dir filepath
+     */
     private String generateOutputFilepath() {
         LocalDateTime now = LocalDateTime.now();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH-mm-ss");
-        String datetime = now.format(formatter);
-        datetime += "-" + String.format("%03d", now.getNano() / 1_000_000);
 
         return new StringBuilder(resultsDir)
-        .append("/" + purpose)
-        .append("/" + datetime)
-        .toString();
+            .append("/" + purpose)
+            .append("/" + now.format(formatter))
+            .append("-" + String.format("%03d", now.getNano() / 1_000_000))
+            .toString();
     }
 
     public String getRecordType() {
