+++ src/main/java/org/example/utils/Utils.java
@@ -1,8 +1,5 @@
 package org.example.utils;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
 /**
  * Defines common helper functions.
  */
