+++ src/main/java/org/example/utils/parsers/ConfigFileParser.java
@@ -9,7 +9,19 @@ import java.util.List;
 
 import org.example.Constants;
 
+/**
+ * Parser for configuration files.
+ */
 public class ConfigFileParser {
+    /**
+     * Gets all lines from the specified configuration file (ignoring comments
+     * and/or blank lines).
+     * 
+     * @param configFilepath
+     * @return all non-blank/non-comment lines
+     * @throws FileNotFoundException
+     * @throws IOException
+     */
     public static List<String> getAllLines(String configFilepath) throws FileNotFoundException, IOException {
         List<String> lines = new ArrayList<>();
